+++ landingzones/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import url
 
-from . import views
+from landingzones import views
 
 app_name = 'landingzones'